Refactor separate client and request handler with owned connections #17
